aded labels and consts to the disasm listing

+++ SharedSource/Utils/BaseWindow.cpp
@@ -25,5 +25,5 @@ void dev::BaseWindow::SetWindowDefaultPosSize()
 {
 	auto windowPos = ImGui::GetWindowPos();
 	ImGui::SetNextWindowPos(ImVec2(windowPos.x, windowPos.y), ImGuiCond_FirstUseEver);
-	ImGui::SetNextWindowSize(ImVec2(m_defaultW, m_defaultH), ImGuiCond_FirstUseEver);
+	ImGui::SetNextWindowSize(ImVec2((float)m_defaultW, (float)m_defaultH), ImGuiCond_FirstUseEver);
 }
